feat(db): update quantity type to Integer and add initial data script

+++ api/src/main/java/com/project/api/dto/recipe/RecipeItemResponseDTO.java
@@ -8,7 +8,7 @@ public record RecipeItemResponseDTO(
         UUID rawMaterialId,
         String rawMaterialName,
         String rawMaterialCode,
-        Double quantity
+        Integer quantity
 ) {
     public static RecipeItemResponseDTO fromEntity(RecipeItem item) {
         return new RecipeItemResponseDTO(
